remote-deployment/requirements.txt
# Lambda-specific dependencies
awslabs.mcp-lambda-handler>=0.1.0
mcp>=1.0.0
boto3>=1.28.0
botocore>=1.31.0
psutil>=5.8.0
